+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Lunaris
 
 ## 1.14.0
+*2019-11-04*
 - Features:
   - Add `lunaris.createUrl(store, method, primaryKey)` method to generate url for a store.
   - Update kitten-format to 1.6.3
-- Fix:
 
 ## 1.13.0
 *2019-09-30*