+++ CHANGELOG.md
@@ -51,6 +51,7 @@
   - When pushing offline transaction, the plugin state was never updated.
   - The method `lunaris._removeAllHooks` was not correctly removing hooks.
   - The internal method `getPrimaryKeyValue()` was trying to cast primary keys as Number.
+  - IndexedDB could block the application initialisation when opening a transaction of an inexistant store.
 
 ## 1.13.0
 *2019-09-30*